[#601] Move smart mix overlay into a component
